+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.1](https://github.com/medly/medly-components/compare/@medly-components/theme@2.10.0...@medly-components/theme@2.10.1) (2021-08-13)
+
+**Note:** Version bump only for package @medly-components/theme
+
+
+
+
+
+# [2.10.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.9.0...@medly-components/theme@2.10.0) (2021-08-11)
+
+
+### Features
+
+* updating theme to include searchBox expandIcon active state ([#379](https://github.com/medly/medly-components/issues/379)) ([f868b93](https://github.com/medly/medly-components/commit/f868b93847ef0ac92f4996897fe3638d5a0b06df))
+
+
+
+
+
+# [2.9.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.8.0...@medly-components/theme@2.9.0) (2021-08-11)
+
+
+### Features
+
+* search box component updates ([#371](https://github.com/medly/medly-components/issues/371)) ([f89ed95](https://github.com/medly/medly-components/commit/f89ed95f9619707aaa47410c99ee83305261473e))
+
+
+
+
+
 # [2.8.0](https://github.com/medly/medly-components/compare/@medly-components/theme@2.7.0...@medly-components/theme@2.8.0) (2021-08-05)